Feature Change: Save Edit when User Clicks off:
Now when user clicks off the edit input, it will no longer close and save the value or revert to previous value.
Instead it will remain open unless as user attempts to edit a different task or add a new task,
if a user performs either of those actions, it will save the value back to text, and focus on the new taks to edit or the add task option.

TODO: Add buttons to save task, and mark as complete.